+++ test/intern.base-config.js
@@ -123,6 +123,7 @@ define([
       'test/unit/when.key.test',
       'test/unit/when.visible-area.test',
       'test/unit/core.worker.test',
+      'test/unit/browser-focusable.test',
     ],
 
     // Functional test suite(s) to run in each browser once non-functional tests are completed