+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The following lists show the changes to the library grouped by domain.
 
 * Adding full support for Internet Explorer 9 - [issue #71](https://github.com/medialize/ally.js/issues/71)
 * Adding full support for Opera 34 (Blink based, behaves like Chrome)
+* Adding manual focusable tests for Safari 6 and 8 on OSX
 * Dropping manual focusable tests for Safari on iOS 8, keeping Safari on iOS 9
 * Dropping manual focusable tests for Mobile Chrome on Android 4.4, keeping Mobile Chrome on Android 5.1
 
@@ -55,6 +56,8 @@ The following lists show the changes to the library grouped by domain.
 * changing modules to be able to load in non-browser environments - [issue #92](https://github.com/medialize/ally.js/issues/92)
 * changing user agent sniffing from detecting browser to rendering engine - [issue #97](https://github.com/medialize/ally.js/issues/97)
 * refactoring `is/is.util.js` to extract image map related functions into `utils/image-map`
+* refactoring `is/focus-relevant` and `is/tabbable` to allow running the identification with execptions via `is/focus-relevant.rules` and `is/tabbable.rules`, while maintaining module signature
+* refactoring `console.log()` to go through `util/logger`
 
 #### Testing